// src/services/rental-api/generated/model/package-item-search-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * pl-car-rental-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PackageItemSearchOutput
 */
export interface PackageItemSearchOutput {
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'packageCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'packageName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'name'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PackageItemSearchOutput
     */
    'price'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof PackageItemSearchOutput
     */
    'minimumDay'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'type'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'imageUrl'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof PackageItemSearchOutput
     */
    'isActive'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'tenantCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PackageItemSearchOutput
     */
    'branchCode'?: string | null;
}